refactor - explore components reused
