--- src/components/Explore/explore.module.css
@@ -1,12 +1,0 @@
-.feedDiv {
-  grid-area: feed;
-  overflow-y: auto;
-  height: 100vh;
-  border-left: solid thin var(--clr-off-primary);
-}
-
-.postDiv {
-  border-bottom: solid thin var(--clr-off-primary);
-  letter-spacing: 1px;
-  line-height: 20px;
-}
