--- src/components/Explore/index.js
@@ -1,1 +1,0 @@
-export { default as ExplorePost } from './ExplorePost'
