+++ src/components/Reusable/index.js
@@ -1,3 +1,4 @@
 export { default as NavBar } from './NavBar'
 export { default as PostsWrapper } from './Posts'
 export { default as FeedPageWrapper } from './FeedPage'
+export { default as Post } from './Post'
