+++ src/components/Reusable/reusable.module.css
@@ -9,9 +9,22 @@
   height: 100vh;
 }
 
+.feedDiv {
+  grid-area: feed;
+  overflow-y: auto;
+  height: 100vh;
+  border-left: solid thin var(--clr-off-primary);
+}
+
 .extraDiv {
   grid-area: extra;
   height: 100vh;
+}
+
+.postDiv {
+  border-bottom: solid thin var(--clr-off-primary);
+  letter-spacing: 1px;
+  line-height: 20px;
 }
 
 @media only screen and (max-width: 1280px) {
